changed the website to react
