+++ frontend/src/app/user/TermsAndCondition/styles.module.css
