+++ frontend/vite.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+export default defineConfig({
+  plugins: [react()],
+  server: {
+    open: true,
+    port: 5173
+  },
+  resolve: {
+    alias: {
+      "@": "/src"
+    }
+  }
+});
